Рефакторинг

+++ src/app/events-dialog/events-dialog.component.ts
@@ -1,16 +1,15 @@
-import { Component, OnInit, Inject, Input} from '@angular/core';
-import {MatDialogRef, MAT_DIALOG_DATA} from '@angular/material/dialog';
-import { EventDataService } from '../services/event-data.service';
-import {Event} from '../classes/event';
-import { Type } from '../classes/type';
+import { Component, OnInit, Inject, Input } from "@angular/core";
+import { MatDialogRef, MAT_DIALOG_DATA } from "@angular/material/dialog";
+import { EventDataService } from "../services/event-data.service";
+import { Event } from "../classes/event";
+import { Type } from "../classes/type";
 
 @Component({
-  selector: 'app-events-dialog',
-  templateUrl: './events-dialog.component.html',
-  styleUrls: ['./events-dialog.component.scss']
+  selector: "app-events-dialog",
+  templateUrl: "./events-dialog.component.html",
+  styleUrls: ["./events-dialog.component.scss"]
 })
 export class EventsDialogComponent implements OnInit {
-
   public title: string = this.data.title || "";
   public type: string = this.data.type || "";
   public description: string = this.data.description || "";
@@ -20,42 +19,63 @@ export class EventsDialogComponent implements OnInit {
   private types: Type[] = [];
   private error: boolean;
 
-  constructor(private eventDataService: EventDataService, 
-              public dialogRef: MatDialogRef<EventsDialogComponent>,
-              @Inject(MAT_DIALOG_DATA) public data: Event) {}
+  constructor(
+    private eventDataService: EventDataService,
+    public dialogRef: MatDialogRef<EventsDialogComponent>,
+    @Inject(MAT_DIALOG_DATA) public data: Event
+  ) {}
 
   ngOnInit() {
     this.events = this.eventDataService.getDataEvent();
     this.types = this.eventDataService.getDataType();
   }
 
-
-  onNoClick(): void {
+  onNoClick() {
     this.dialogRef.close();
   }
 
-  onAddClick(): void {
+  onAddClick() {
     this.error = false;
-    if((this.title.length > 1 || this.title.length < 11) && this.type != "" && this.date != "" && this.time != ""){
-      let newDate = new Date (this.date);
-      let newMonth: string = ((newDate.getMonth() + 1) < 10) ? `0${newDate.getMonth() + 1}` : `${newDate.getMonth() + 1}`;
-      let newDay: string = (newDate.getDate() < 10) ? `0${newDate.getDate()}` : `${newDate.getDate()}`;
-      let updatedDate: string = `${newDate.getFullYear()}-${newMonth}-${newDay}`;
+    if (
+      this.title.length > 1 &&
+      this.title.length < 11 &&
+      this.type !== "" &&
+      this.date !== "" &&
+      this.time !== ""
+    ) {
+      const newDate = new Date(this.date);
+      let updatedDate: string = this.eventDataService.formattingOfTheDate(
+        newDate
+      );
 
-      if(this.data.date == updatedDate && this.data.time == this.time){
-        this.eventDataService.removeEvent(this.data.date, this.data.time)
-        this.eventDataService.addEvent(updatedDate, this.time, this.title, this.description, this.type);
+      if (this.data.date === updatedDate && this.data.time === this.time) {
+        this.eventDataService.removeEvent(this.data.date, this.data.time);
+        this.eventDataService.addEvent(
+          updatedDate,
+          this.time,
+          this.title,
+          this.description,
+          this.type
+        );
         this.dialogRef.close();
-      }
-      else{
-        if(!this.events.find(item => ((item.date == updatedDate) && (item.time == this.time)))){
-          if(this.data.date != undefined){
-            this.eventDataService.removeEvent(this.data.date, this.data.time)
+      } else {
+        if (
+          !this.events.find(
+            item => item.date === updatedDate && item.time === this.time
+          )
+        ) {
+          if (this.data.date !== undefined) {
+            this.eventDataService.removeEvent(this.data.date, this.data.time);
           }
-          this.eventDataService.addEvent(updatedDate, this.time, this.title, this.description, this.type);
+          this.eventDataService.addEvent(
+            updatedDate,
+            this.time,
+            this.title,
+            this.description,
+            this.type
+          );
           this.dialogRef.close();
-        }
-        else {
+        } else {
           this.error = true;
         }
       }
